[JOBCTL] Finished jobs/fg/bg

+++ srcs/module/interpreter/jobcontrol/jobctl.c
@@ -6,42 +6,104 @@
 /*   By: skuppers <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/06/20 18:17:58 by skuppers          #+#    #+#             */
-/*   Updated: 2019/06/24 19:37:01 by skuppers         ###   ########.fr       */
+/*   Updated: 2019/06/25 16:00:35 by skuppers         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 # include "sh21.h"
 
+static void			remove_job_from_active_list(t_list **list, t_job *job)
+{
+	t_list	*ptr;
+	t_list	*job_ptr;
+
+	if (job == NULL)
+		return ;
+	job_ptr = *list;
+
+	while (job_ptr != NULL && (t_job*)job_ptr->data != job)
+		job_ptr = job_ptr->next;
+
+	if (job_ptr == *list)
+		*list = (job_ptr->next);
+	else
+	{
+		ptr = *list;
+		while (ptr->next != NULL && ptr->next != job_ptr)
+			ptr = ptr->next;
+		ptr->next = ptr->next->next;
+	}
+}
+
 static void			job_to_registry(t_registry *shell, t_job *job)
 {
+	char		*avs;
 	t_list		*data;
 	t_job		job_cpy;
 
+	avs = NULL;
 	ft_bzero(&job_cpy, sizeof(t_job));
 	ft_memcpy(&job_cpy, job, sizeof(t_job));
+
+	get_job_av(job, &avs);
+	ft_printf("[%d]+  Stopped(%d) \t %s\n", job->id,
+					job->signo, avs);
+	ft_strdel(&avs);
+
 	job->processes = NULL;
 	job->term_modes = NULL;
 	data = ft_lstnew(&job_cpy, sizeof(t_job));
 	ft_lstaddback(&shell->job_list, data);
 	shell->current_minus = shell->current_plus;
 	shell->current_plus = data;
+
 }
 
+void	update_job_ids(t_registry *shell)
+{
+	uint32_t	id;
+	t_job		*job;
+	t_list		*job_ptr;
+
+	id = 1;
+	job_ptr = shell->job_list;
+	while (id <= shell->active_jobs && job_ptr != NULL)
+	{
+		job = job_ptr->data;
+		job->id = id;
+		++id;
+		job_ptr = job_ptr->next;
+	}
+}
 
 void	job_to_foreground(t_registry *shell, t_job *job)
 {
+	char	*avs;
+
+	if (job == NULL)
+		return ;
 	job->state = RUNNING;
-//	shell->active_jobs--;
-
-//	ft_lstdel(&,);
 	((t_process*)job->processes->data)->stopped = FALSE;
-
-//	ft_dprintf("Attaching pid %d to terminal\n", job->pgid);
-
+	remove_job_from_active_list(&shell->job_list, job);
+	shell->active_jobs--;
+	update_job_ids(shell);
+	get_job_av(job, &avs);
+	ft_printf("%s\n", avs);
+	ft_strdel(&avs);
 	tcsetpgrp(STDOUT_FILENO, job->pgid);
-
 	killpg(job->pgid, SIGCONT);
 	waiter(shell, job);
+}
+
+void	job_run_background(__unused t_registry *shell, t_job *job)
+{
+	if (job->state == STOPPED)
+	{
+		job->state = RUNNING;
+		job->signo = SIGCONT;
+//		ft_printf("Resuming [%d] in background.\n", job->pgid);
+		killpg(job->pgid, SIGCONT);
+	}
 }
 
 void	jobctl(__unused t_registry *shell, t_job *job, uint8_t flag)
@@ -54,4 +116,8 @@ void	jobctl(__unused t_registry *shell, t_job *job, uint8_t flag)
 
 	else if (flag & JOBCTL_PUTINFG)
 		job_to_foreground(shell, job);
+
+	else if (flag & JOBCTL_RUNINBG)
+		job_run_background(shell, job);
+
 }
